python3Packages.dateparser: patch flaky test
The dateparser module has a test that works with the current day of the
month and uses it in June 2020. This breaks everytime we build the
package on the 31st of a month, because June only has 30 days.

```
❯ cal 6 2020
      Juni 2020
Mo Di Mi Do Fr Sa So
 1  2  3  4  5  6  7
 8  9 10 11 12 13 14
15 16 17 18 19 20 21
22 23 24 25 26 27 28
29 30
```

https://github.com/scrapinghub/dateparser/issues/1053
(cherry picked from commit 0cece84ef3b7ae1a2cd00b3f9893673e1b7e45e2)