nixos/no-x-libs: remove X libs from ffmpeg
Prior to this change, ffmpeg couldn't be built for an
environment.noXlibs system, because it would fail in:

	ffmpeg → SDL2 → libdecor

ffmpeg certainly does not need support for SDL2 windowing on a noXlibs
system.

This fix is important because the minidlna NixOS test, which uses the
minimal profile (and therefore environment.noXlibs) and ffmpeg, can't
currently build.