bintools-wrapper: move gcc LTO plugin population from binutils
Note: it's an incompletel wrap: we only populate plugin link for
`ld`, `strip` (existing wrappers) and `ar` (new wrapper).

If we go that route we'll need to wrap most binaries for GNU binutils
as almost all of them can detect plugin format and act on it accordingly:
`size`, `nm`, `ranlib`, `addr2line`.