nixos/tests/cagebreak: Enable just polkit instead of udisks2
The udisks2 service was enabled to fix the test in (c5ebec7ee42).
However, cagebreak doesn't require udisks2, just polkit (which the
udisks2 module enables and which is why the cagebreak test broke after
the udisks2 module was disabled by default).

I've documented why polkit is required in this PR:
https://github.com/NixOS/nixpkgs/pull/156858

In this case the "dependency" chain is basically cagebreak -> wlroots ->
libseat -> logind (with polkit support) -> polkit.