paperless: Allow passing extraPropagatedBuildInputs
The paperless package exposes a PYTHONPATH via passthru, that is used
from its accompanying module. Using `overridePythonAttrs` it is not
possible to affect the passed PYTHONPATH to inject additional
dependencies.

My specific use case is to patch Django's settings.py with LDAP support
and propagate the `django-auth-ldap` package, which overall has a low
maintenance cost.