paperless: move `PYTHONPATH` definition to module
`paperless-ngx.pythonPath` was incomplete due to the missing paperless-ngx
source, so it had to be amended in the service.
Instead of amending it, define it entirely in the service.

This allows an override of `paperless-ngx.propagatedBuildInputs` to be reflected
in the service's PYTHONPATH.

(cherry picked from commit 783f8f16c16ba47daffff2c6b92a96eb8bb5d363)