nixos/kanidm: Bind mount cacert path in unixd service
In order to be able to use the unixd service with the `verify_ca` and
`verify_hostnames` set to `true` it needs to be able to read the
certificate store. This change bind mounts the cacert paths for the
unixd service.

(cherry picked from commit 3df41451e3f5179e1d02cf8366f1646ff3eb94ae)