p4v: use autoPatchelfHook, upstream libraries
P4V comes with a compatible set of Qt libraries that just need to be
patchelf'd to be usable, which avoids the need for the crypto hack we
were using. Switch to using autoPatchelfHook for all libraries, and
prefer the distribution's Qt libraries as part of that.

This also fixes a bug in the distributed version of p4d, which prevented
DVCS from working correctly in P4V.