fix: fetch packages in buildPhase
The reason for that is that sometimes you may need to run additional actions to compile dependencies.

For example when there are Rustler dependencies then you need write access to the `deps/` directory.
With current implementation it is not possible to run any steps after deps are fetched but before these
are installed. This change allows user to use `postBuild` step to contain such actions. Additionally it adds
support for providing custom `nativeBuildInputs` that may be needed for above step.