nextcloud: drop password regeneration behavior
While updating to 23.0.9/24.0.5[1], it was discovered that Nextcloud
silently changes db passwords in some cases (in case of MySQL in **all**
now).

This is inherently incompatible with our module. Further context is
provided in the patch file attached to this commit.

[1] https://github.com/NixOS/nixpkgs/pull/190646, see comments below

(cherry picked from commit 958914fab2e9754b8e17147457515f73a7d8c287)