fetchCargoTarball: dont require setting hash/sha256
It used to be required to set one of the flags influencing the FOD
output hash before we would allow a derivation to evaluate. This means
the user would usually need to set hash/sha256 to lib.fakeHash or an
empty string, otherwise we would throw.

This is inconsistent with other FODs such as fetchFromGitHub, which
allow you to neglect setting the hash. This will print a warning when
evaluation starts, and dumps the correct hash after a mismatch is
inevitably encountered.

This allows you to create an expression that will "prefetch" your
projects dependencies with less boilerplate, saving some typing.