buildRustPackage: dont throw if cargoHash/cargoSha256 are not set
This removes an assert that checked if one of the FOD output hash
attributes were set, in the case cargoVendorDir/cargoLock were not.

If neither of those are used fetchCargoTarball is invoked to generate a
FOD, which now allows you to not pass it any hash so it can "prefetch"
the hash.

A warning will get printed explaining an empty hash is found, and
the fetcher will begin to do its magic. Afterwards the fetcher will
inevitably fail with a hash mismatch, and you can copy paste the
correct hash into either one of FOD output hash attributes.

This allows you to create an expression that will generate your
derivations hashes with less boilerplate, saving some typing.