sublime-merge: Avoid absolute exec paths in desktop file
When user installed a new version of SM (or just a different derivation),
GNOME Shell would lose the association of the windows with the icon.
Since the package needs to be installed in order for the desktop file
to be available, having the path absolute is useless any way.

Same change as the one for sublime4 in the parent commit.