appimageTools.wrapType2: add default install commands and extraInstall
It used to be common for derivations using 'wrapType2' to manually
extract the contents of appimage, just so they could reference it
in 'extraInstallCommands'.

With that the risk of duplicated store paths increases, as 'wrapType2'
would call the same function the extract the appimage but with a
potential difference in arguments. It is also just needless boiletplate.

To fix this problem 'extraInstall' is introduced, which has a reference
to the extracted appimage as function argument. Much like
'buildFhsUserEnv's 'targetPkgs' does.

Default 'extraInstallCommands' are also added to automate stuff pretty
much all appimage-based derivations do manually right now:

* Remove the derivations version from the path of the installed binary.
* Install the desktop file, which is required to be bundled with any
  appimage.