haskellPackages.callCabal2nixWithOptions: allow setting source filter
This is useful if you use hpack's ability to include files: previously
callCabal2nix did not allow you to include those extra yaml files.

This change is backward compatible: it introduces a
callCabal2nixWithOptions' that has the new features as an attribute set,
then changes callCabal2nixWithOptions to work as previously, but as a
wrapper of callCabal2nixWithOptions'.