xcaddy: use temp dir on Darwin
Upstream xcaddy uses the $HOME directory as the temp directory on Darwin
in order to fix an interesting bug.

In order to find version information for Caddy, it runs a command that
needs to operate on file system without "noexec" set, but Darwin's temp
directory has it set.

We can tolerate losing version information, since the alternative is
that xcaddy does not run on Darwin.