subversion: openssl_3 -> openssl_1_1 to fix https lock
Subversion hangs when accessing an HTTPS repository.
Rollback to openssl_1_1 fixes the problem.

Setting `openssl = openssl_1_1` only for subversion is not enough.
I had to also set it for the `aprutil` instance used by the `serf`
library used by `subversion`. I do not know if this is a distinct bug.

I first reported the problem at
https://discourse.nixos.org/t/how-to-find-what-has-changed-between-two-builds-of-a-derivation/21854?u=mnuccioarpae