linux/kernel: backport symbol GPLification fix from 5.19
Linux has a few PowerPC-specific symbols which are marked as GPL exports; these
symbols wound up being exposed in Linux 5.12 and are needed by OpenZFS.  The
symbol licensing was fixed in mainline 5.19; this commit backports the fix to
all previous affected kernels.

This commit is required in order to build the NixOS ISO for PowerPC64.