asciidoc: pass `SOURCE_DATE_EPOCH` to xsltproc
Because we override ENV, we reset the environment when invoking xsltproc.

This is a good thing, since it also makes it much faster and less dependent
on the environment - but in case of `SOURCE_DATE_EPOCH` we actually want
this environment variable to be passed on, to make sure `a2x` does not leak
the build date into the generated manpages when this is not desired.

Fixes #192726