ntfy: restructure to be a Python package
This allows us to access ntfy built with a specific Python interpreter
more easily (and thus swapping the Python interpreter uses).

This also splits out the things which are ntfy optional dependencies so
that they can be disabled - in particular, sleekxmpp is broken on Darwin
and because of the way the package is designed at the moment it isn't
possible to turn it off.

e.g. this allows you to use ntfy with python3.10 as long as you disable
XMPP support.