ntfy: add options to turn off optional features
This splits out the things which are ntfy optional dependencies so
that they can be disabled - in particular, sleekxmpp is broken on Darwin
and because of the way the package is designed at the moment it isn't
possible to turn it off.

This also allows you to use newer or older python interpreters without
overriding e.g. "python39" to actually mean python310.