kubectl: add Krew plugins
The Krew plug-in manager for kubectl tracks a set of available kubectl plug-ins.
We automatically generate derivations based on plug-in manifests from the Krew
plug-in repository. We can then make plug-ins available to kubectl using the
“withKrewPlugins” attribute of the kubectl derivation, e.g.:

    kubectl.withKrewPlugins (k: [ k.node-shell ])

Most plug-ins are statically compiled binaries. Some may have additional
requirements regarding commands available on the system path, which we cannot
handle automatically.