ruby: Make version overridable
Previously, because the `version` attrset passed to `generic` was not
overridable, trying to use this to change the Ruby version to, e.g., a
different point release like so:

```
final: prev: {
  ruby_3_1_1 = prev.ruby_3_1.overrideAttrs (old: {
    version = "3.1.1";
    src = fetchurl { ... };
  });
}
```

would result in some unexpected behavior as the original version would
still be used. For example, the shared library built would be for the
*original* version, not the overridden one, and so have a confusing
name like `libruby-3.1.2.so.3.1.1`.

This moves the args to `generic` into the `makeOverridable` function,
allowing something like this:

```
final: prev: {
  ruby_3_1_1 = ruby_3_1.override {
    version = let
      rubyVersion = import ../development/interpreters/ruby/ruby-version.nix { inherit lib; };
    in rubyVersion "3" "1" "1" "";
    sha256 = "...";
    patchsetsVersion = ruby_3_1.passthru.version; # or = null
  };
}
```

As `patchsets.nix` in nixpkgs uses specific versions, optionally allow
using patchsets from a different version than `versions` to, e.g., use
those of a similar release. Or just specify `null` to ignore them.

This is helpful for some downstream consumers as they may get a specific
version they need without needing to patch nixpkgs, pin old nixpkgs, or
maintain out-of-tree copies of the derivations. As nixpkgs is not
explicitly supporting different versions, the burden is still largely on
downstream. Meanwhile, it may help encourage adoption by providing more
flexibility to downstreams that need it.