kaiteki: init at 2022-09-03
kaiteki is a Mastodon/Pleroma/Misskey ActivityPub client.

Pleroma and Mastodon support are verified working. tested:
- logging in
- browsing the timeline

for file uploading, Kaiteki tests for a few different file-choosers that
may be installed on the system, at runtime. for a gnome system, the user
will want to place `pkgs.gnome.zenity` into the environment for that
feature to work.

upstream doesn't use a versioning scheme yet, hence the date-based
version chosen in this package.

there doesn't appear to be a `.packages` file, which the nixpkgs
`mkFlutterApp` expects (for flutter 3). i don't know much about this: i
just made the code that touches it conditional and it seems to work.