matrix-synapse: 1.67.0 -> 1.68.0
https://github.com/matrix-org/synapse/releases/tag/v1.68.0

Uses poetry-core to build the package. Drops setuptools-rust from
runtime dependencies, because they are checked at startup, which breaks
because we do a clear separation of concerns.

Fixes a misconception about the tests, that were until now always run
against the source module. This breaks with the introduction of the rust
components, because they're not available in the source module.

Adds missing pre and post hooks to checkPhase.