go: do not remove regexp/syntax/make_perl_groups.pl
This change keeps the regexp/syntax/make_perl_groups.pl file even though
it contains invalid shebang. It is not used by the test suite and there
is no reason to remove it. In fact, it is confusing when some files used
to generate Go code in standard library are missing in the package.