Update source to fix #194116
Before this commit, paths to GUI config and manual were written to the config file. When running gc after an update, these paths do not exist anymore.
That changes with this commit: Paths are commented out in config file. Nonexisting paths are removed from config file when running beamerpresenter.

Other changes in the code compared to v0.2.3: compatibility with Qt 6.4