nixos/nixpkgs: make `nixpkgs.config` mergeable
Closes #194056
Alternative implementation to #80582

`nixpkgs.config` is now a freeform-type which imports the
`<nixpkgs/pkgs/top-level/config.nix>`-module. This brings the following
benefits:

* Proper merging (to a certain degree, more details later) of several
  `nixpkgs.config` declarations.

* `allowUnfree` and friends appear in the NixOS manual.

To be precise, this means that something like this is now possible:

    {
      nixpkgs = mkMerge [
        { allowUnfree = false; }
        { allowUnfree = mkForce true; }
      ];
    }

Previously this would lead to a bogus attr-set, namely

    { _type = "override"; content = true; priority = 50; }

since no merging for this option-type was implemented.

This however has certain limitations. For instance

    nixpkgs.config.allowUnfree = mkForce false;

works fine whereas

    nixpkgs.config.virtualbox.enableExtensionPack = mkForce true;

doesn't. This is because we use `types.raw` inside the config-module for
`nixpkgs` which doesn't merge undeclared attribute-sets (such as
`virtualbox`). This is because we don't know if any attribute-set is
actually mergeable, for further details see the previous discussion[1].

A basic type has been implemented for the (actually deprecated)
mechanisms `packageOverrides`/`perlPackageOverrides`. This one basically
applies the `pkgs` argument onto each definition of this function, so
e.g. the following expression

    {
      nixpkgs.config = mkMerge [
        { packageOverrides = pkgs: { randomattr = 23; }; }
        { packageOverrides = pkgs: { randomattr = pkgs.randomattr + 19; }; }
      ];
    }

adds an attribute `randomattr` of value `42` to `_module.args.pkgs`.

[1] https://github.com/NixOS/nixpkgs/pull/194207#issuecomment-1267002199